Revert "final mediump"
This reverts commit d78b7e731f8b5b533ad7859663a8f2664d28f3bc.

+++ src/webl/shaders/shaderBackgroundScene.ts
@@ -107,8 +107,8 @@ export function GetShaderSourceBackgroundFloorRenderPerspectiveVS() {
 export function GetShaderSourceBackgroundFloorRenderPerspectivePS() {
     return /* glsl */ `#version 300 es
 		
-		precision mediump float;
-		precision mediump sampler2D;
+	precision mediump float;
+	precision mediump sampler2D;
 	
 		layout(location = 0) out vec4 OutColor;
 	
@@ -899,8 +899,8 @@ export function GetShaderSourceSpotlightRenderVS() {
 export function GetShaderSourceSpotlightRenderPS() {
     return /* glsl */ `#version 300 es
 	
-	precision mediump float;
-	precision mediump sampler2D;
+	precision highp float;
+	precision highp sampler2D;
 
 	layout(location = 0) out float outSpotlightColor;
